+++ xpaw/config.py
@@ -4,7 +4,7 @@ import copy
 from collections import MutableMapping
 import types
 
-from . import defaultconfig
+from xpaw import defaultconfig
 
 CONFIG_PRIORITIES = {
     "default": 0,